+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Bug fixes:
+  * Fixed compilation issue in C++ caused by redundant "const" modifiers on methods.
+
 ## 7.1.4
 Release date: 2020-07-06
 ### Features: